#13804: support test utils: get_lib_dtype
